Major revision: Multiple updates and feature adds.
Added support for JSON, which added two new options; "bJson" and "oData', "bJson" is a Boolean option that is either true or false, and the "oData" is expecting an object, example [{"weight":"","value":""}]. Removed the need to specify the "<a>" tag in the jQuery selector, which improves performance with .find(); however,the option "sSelector" was added which allows you to select the "<a>" option if desired, or to specify any other element. In addition, the pull request for rel vs data-weight was resolved by adding the "sWeightSelector" which will change allow you to specify which to use; data-weight="" or rel="". All the options are being verify against the correct type, and defaults are set within the "oOptions" variable.